backend/backend_foodgramm/requirements.txt
asgiref==3.8.1
asttokens==3.0.0
atomicwrites==1.4.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==2.0.12
colorama==0.4.6
cryptography==44.0.0
decorator==5.1.1
defusedxml==0.8.0rc2
Django==3.2
django-filter==23.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
exceptiongroup==1.2.2
executing==2.1.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
ipython==8.18.1
jedi==0.19.2
matplotlib-inline==0.1.7
oauthlib==3.2.2
packaging==24.2
parso==0.8.4
pillow==11.0.0
pluggy==0.13.1
prompt_toolkit==3.0.48
psycopg2==2.9.10
pure_eval==0.2.3
py==1.11.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
requests==2.26.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
stack-data==0.6.3
toml==0.10.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==1.26.20
wcwidth==0.2.13
